(CODEX) Replace admin router with service layer
- add AdminDatastoreService with merge + cache orchestration so routes depend on service abstraction
- replace legacy bt_servant_engine.apps.api.routes.admin with new admin_datastore router and adjust dependencies/docs/tests accordingly
- move merge helpers into services.admin package to satisfy onion layering and update unit tests for new locations and typing

Test Plan:
- scripts/check_repo.sh
