+++ bt_servant_engine/apps/api/routes/__init__.py
@@ -1,5 +1,5 @@
 """Router namespace exports for FastAPI include hooks."""
 
-from . import admin, health, webhooks
+from . import admin_datastore, admin_logs, health, webhooks
 
-__all__ = ["admin", "health", "webhooks"]
+__all__ = ["admin_datastore", "admin_logs", "health", "webhooks"]
