+++ bt_servant_engine/services/admin/__init__.py
@@ -1,0 +1,5 @@
+"""Administrative service layer utilities for datastore management."""
+
+from .datastore import AdminDatastoreService
+
+__all__ = ["AdminDatastoreService"]
